llama restructuring
