--- src/monocle_apptrace/constants.py
@@ -1,36 +1,0 @@
-# Azure environment constants
-AZURE_ML_ENDPOINT_ENV_NAME = "AZUREML_ENTRY_SCRIPT"
-AZURE_FUNCTION_WORKER_ENV_NAME = "FUNCTIONS_WORKER_RUNTIME"
-AZURE_APP_SERVICE_ENV_NAME = "WEBSITE_SITE_NAME"
-AWS_LAMBDA_ENV_NAME = "AWS_LAMBDA_RUNTIME_API"
-GITHUB_CODESPACE_ENV_NAME = "CODESPACES"
-
-AWS_LAMBDA_FUNCTION_IDENTIFIER_ENV_NAME = "AWS_LAMBDA_FUNCTION_NAME"
-AZURE_FUNCTION_IDENTIFIER_ENV_NAME = "WEBSITE_SITE_NAME"
-AZURE_APP_SERVICE_IDENTIFIER_ENV_NAME = "WEBSITE_DEPLOYMENT_ID"
-GITHUB_CODESPACE_IDENTIFIER_ENV_NAME = "GITHUB_REPOSITORY"
-# Azure naming reference can be found here
-# https://learn.microsoft.com/en-us/azure/cloud-adoption-framework/ready/azure-best-practices/resource-abbreviations
-AZURE_FUNCTION_NAME = "azure.func"
-AZURE_APP_SERVICE_NAME = "azure.asp"
-AZURE_ML_SERVICE_NAME = "azure.mlw"
-AWS_LAMBDA_SERVICE_NAME = "aws.lambda"
-GITHUB_CODESPACE_SERVICE_NAME = "github_codespace"
-
-# Env variables to identify infra service type
-service_type_map = {
-    AZURE_ML_ENDPOINT_ENV_NAME: AZURE_ML_SERVICE_NAME,
-    AZURE_APP_SERVICE_ENV_NAME: AZURE_APP_SERVICE_NAME,
-    AZURE_FUNCTION_WORKER_ENV_NAME: AZURE_FUNCTION_NAME,
-    AWS_LAMBDA_ENV_NAME: AWS_LAMBDA_SERVICE_NAME,
-    GITHUB_CODESPACE_ENV_NAME: GITHUB_CODESPACE_SERVICE_NAME
-}
-
-# Env variables to identify infra service name
-service_name_map = {
-    AZURE_APP_SERVICE_NAME: AZURE_APP_SERVICE_IDENTIFIER_ENV_NAME,
-    AZURE_FUNCTION_NAME: AZURE_FUNCTION_IDENTIFIER_ENV_NAME,
-    AZURE_ML_SERVICE_NAME: AZURE_ML_ENDPOINT_ENV_NAME,
-    AWS_LAMBDA_SERVICE_NAME: AWS_LAMBDA_FUNCTION_IDENTIFIER_ENV_NAME,
-    GITHUB_CODESPACE_SERVICE_NAME: GITHUB_CODESPACE_IDENTIFIER_ENV_NAME
-}
