+++ src/monocle_apptrace/exporters/aws/s3_exporter.py
@@ -24,8 +24,8 @@ class S3SpanExporter(SpanExporterBase):
     def __init__(self, bucket_name=None, region_name=None):
         super().__init__()
         # Use environment variables if credentials are not provided
-        DEFAULT_FILE_PREFIX = "monocle_trace_"
-        DEFAULT_TIME_FORMAT = "%Y-%m-%d_%H.%M.%S"
+        DEFAULT_FILE_PREFIX = "monocle_trace__"
+        DEFAULT_TIME_FORMAT = "%Y-%m-%d__%H.%M.%S"
         self.max_batch_size = 500
         self.export_interval = 1
         self.s3_client = boto3.client(
