+++ src/monocle_apptrace/instrumentation/common/base_task_processor.py
@@ -1,0 +1,11 @@
+from abc import ABC, abstractmethod
+
+class TaskProcessor(ABC):
+
+    @abstractmethod
+    def pre_task_processing(self, to_wrap, wrapped, instance, args, span):
+        pass
+
+    @abstractmethod
+    def post_task_processing(self, to_wrap, wrapped, instance, args, kwargs, return_value, span):
+        pass
