+++ src/monocle_apptrace/instrumentation/metamodel/llamaindex/methods.py
@@ -1,0 +1,99 @@
+from monocle_apptrace.instrumentation.common.wrapper import atask_wrapper, task_wrapper
+from monocle_apptrace.instrumentation.metamodel.llamaindex.inference import (
+    INFERENCE,
+)
+from monocle_apptrace.instrumentation.metamodel.llamaindex.retrieval import (
+    RETRIEVAL,
+)
+
+
+LLAMAINDEX_METHODS = [
+    {
+        "package": "llama_index.core.indices.base_retriever",
+        "object": "BaseRetriever",
+        "method": "retrieve",
+        "span_name": "llamaindex.retrieve",
+        "framework_type": "llama_index",
+        "wrapper_method": task_wrapper,
+        "output_processor": RETRIEVAL
+    },
+    {
+        "package": "llama_index.core.indices.base_retriever",
+        "object": "BaseRetriever",
+        "method": "aretrieve",
+        "span_name": "llamaindex.retrieve",
+        "framework_type": "llama_index",
+        "wrapper_method": atask_wrapper,
+        "output_processor": RETRIEVAL
+    },
+    {
+        "package": "llama_index.core.base.base_query_engine",
+        "object": "BaseQueryEngine",
+        "method": "query",
+        "span_name": "llamaindex.query",
+        "framework_type": "llama_index",
+        "wrapper_method": task_wrapper
+    },
+    {
+        "package": "llama_index.core.base.base_query_engine",
+        "object": "BaseQueryEngine",
+        "method": "aquery",
+        "span_name": "llamaindex.query",
+        "framework_type": "llama_index",
+        "wrapper_method": atask_wrapper
+    },
+    {
+        "package": "llama_index.core.llms.custom",
+        "object": "CustomLLM",
+        "method": "chat",
+        "span_name": "llamaindex.llmchat",
+        "framework_type": "llama_index",
+        "wrapper_method": task_wrapper,
+        "output_processor": INFERENCE
+    },
+    {
+        "package": "llama_index.core.llms.custom",
+        "object": "CustomLLM",
+        "method": "achat",
+        "span_name": "llamaindex.llmchat",
+        "framework_type": "llama_index",
+        "wrapper_method": atask_wrapper,
+        "output_processor": INFERENCE
+    },
+    {
+        "package": "llama_index.llms.openai.base",
+        "object": "OpenAI",
+        "method": "chat",
+        "span_name": "llamaindex.openai",
+        "framework_type": "llama_index",
+        "wrapper_method": task_wrapper,
+        "output_processor": INFERENCE
+    },
+    {
+        "package": "llama_index.llms.openai.base",
+        "object": "OpenAI",
+        "method": "achat",
+        "span_name": "llamaindex.openai",
+        "framework_type": "llama_index",
+        "wrapper_method": atask_wrapper,
+        "output_processor": INFERENCE
+    },
+    {
+        "package": "llama_index.llms.mistralai.base",
+        "object": "MistralAI",
+        "method": "chat",
+        "span_name": "llamaindex.mistralai",
+        "framework_type": "llama_index",
+        "wrapper_method": task_wrapper,
+        "output_processor": INFERENCE
+    },
+    {
+        "package": "llama_index.llms.mistralai.base",
+        "object": "MistralAI",
+        "method": "achat",
+        "span_name": "llamaindex.mistralai",
+        "framework_type": "llama_index",
+        "wrapper_method": atask_wrapper,
+        "output_processor": INFERENCE
+    }
+]
