--- src/monocle_apptrace/metamodel/entities/vector_store_types.json
@@ -1,25 +1,0 @@
-{
-    "entity" : "vector_store",
-    "types": [
-        {
-            "name": "chroma",
-            "attributes" : []
-        },
-        {
-            "name": "aws_es",
-            "attributes" : []
-        },
-        {
-            "name": "milvus",
-            "attributes" : []
-        },
-        {
-            "name": "pinecone",
-            "attributes" : []
-        },
-        {
-            "name": "generic",
-            "attributes" : []
-        }
-    ]
-}
