--- src/monocle_apptrace/metamodel/entities/workflow_types.json
@@ -1,22 +1,0 @@
-{
-    "entity" : "workflow",
-    "types": [
-        {
-            "name": "llama_index",
-            "attributes" : []
-        },
-        {
-            "name": "langchain",
-            "attributes" : []
-        },
-        {
-            "name": "haystack",
-            "attributes" : []
-        },
-        {
-            "name": "generic",
-            "attributes" : []
-        }
-
-    ]
-}
