--- src/monocle_apptrace/metamodel/maps/attributes/retrieval/haystack_entities.json
@@ -1,31 +1,0 @@
-{
-  "type": "retrieval",
-  "attributes": [
-    [
-      {
-        "_comment": "vector store name and type",
-        "attribute": "name",
-         "accessor": "lambda arguments: resolve_from_alias(arguments['instance'].__dict__, ['document_store', '_document_store']).__class__.__name__"
-      },
-      {
-        "attribute": "type",
-        "accessor": "lambda arguments: 'vectorstore.'+resolve_from_alias(arguments['instance'].__dict__, ['document_store', '_document_store']).__class__.__name__"
-      },
-      {
-        "attribute": "deployment",
-        "accessor": "lambda arguments: get_vectorstore_deployment(resolve_from_alias(arguments['instance'].__dict__, ['document_store', '_document_store']).__dict__)"
-      }
-    ],
-    [
-      {
-        "_comment": "embedding model name and type",
-        "attribute": "name",
-        "accessor": "lambda arguments: get_embedding_model()"
-      },
-      {
-        "attribute": "type",
-        "accessor": "lambda arguments: 'model.embedding.'+get_embedding_model()"
-      }
-    ]
-  ]
-}
