--- src/monocle_apptrace/metamodel/maps/attributes/retrieval/langchain_entities.json
@@ -1,31 +1,0 @@
-{
-  "type": "retrieval",
-  "attributes": [
-    [
-      {
-        "_comment": "vector store name and type",
-        "attribute": "name",
-        "accessor": "lambda arguments: type(arguments['instance'].vectorstore).__name__"
-      },
-      {
-        "attribute": "type",
-        "accessor": "lambda arguments: 'vectorstore.'+type(arguments['instance'].vectorstore).__name__"
-      },
-      {
-        "attribute": "deployment",
-        "accessor": "lambda arguments: get_vectorstore_deployment(arguments['instance'].vectorstore.__dict__)"
-      }
-    ],
-    [
-      {
-        "_comment": "embedding model name and type",
-        "attribute": "name",
-        "accessor": "lambda arguments: arguments['instance'].vectorstore.embeddings.model"
-      },
-      {
-        "attribute": "type",
-        "accessor": "lambda arguments: 'model.embedding.'+arguments['instance'].vectorstore.embeddings.model"
-      }
-    ]
-  ]
-}
