--- src/monocle_apptrace/metamodel/maps/attributes/retrieval/llamaindex_entities.json
@@ -1,31 +1,0 @@
-{
-  "type": "retrieval",
-  "attributes": [
-    [
-      {
-        "_comment": "vector store name and type",
-        "attribute": "name",
-        "accessor": "lambda arguments: type(arguments['instance']._vector_store).__name__"
-      },
-      {
-        "attribute": "type",
-        "accessor": "lambda arguments: 'vectorstore.'+type(arguments['instance']._vector_store).__name__"
-      },
-      {
-        "attribute": "deployment",
-        "accessor": "lambda arguments: get_vectorstore_deployment(arguments['instance']._vector_store)"
-      }
-    ],
-    [
-      {
-        "_comment": "embedding model name and type",
-        "attribute": "name",
-        "accessor": "lambda arguments: arguments['instance']._embed_model.model_name"
-      },
-      {
-        "attribute": "type",
-        "accessor": "lambda arguments: 'model.embedding.'+arguments['instance']._embed_model.model_name"
-      }
-    ]
-  ]
-}
