--- src/monocle_apptrace/metamodel/maps/botocore_methods.json
@@ -1,13 +1,0 @@
-{
-  "wrapper_methods": [
-    {
-      "package": "botocore.client",
-      "object": "ClientCreator",
-      "method": "create_client",
-      "wrapper_package": "wrap_common",
-      "wrapper_method": "task_wrapper",
-      "skip_span": true,
-      "output_processor": ["metamodel/maps/attributes/inference/botocore_entities.json"]
-    }
-  ]
-}
