--- src/monocle_apptrace/metamodel/maps/haystack_methods.json
@@ -1,45 +1,0 @@
-{
-"wrapper_methods" :  [
-    {
-        "package": "haystack_integrations.components.retrievers.opensearch",
-        "object": "OpenSearchEmbeddingRetriever",
-        "method": "run",
-        "span_name": "haystack.retriever",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "task_wrapper",
-        "output_processor": ["metamodel/maps/attributes/retrieval/haystack_entities.json"]
-    },
-    {
-        "package": "haystack.components.retrievers.in_memory",
-        "object": "InMemoryEmbeddingRetriever",
-        "method": "run",
-        "span_name": "haystack.retriever",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "task_wrapper",
-        "output_processor": ["metamodel/maps/attributes/retrieval/haystack_entities.json"]
-    },
-    {
-        "package": "haystack.components.generators.openai",
-        "object": "OpenAIGenerator",
-        "method": "run",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "llm_wrapper",
-        "output_processor": ["metamodel/maps/attributes/inference/haystack_entities.json"]
-    },
-    {
-        "package": "haystack.components.generators.chat.openai",
-        "object": "OpenAIChatGenerator",
-        "method": "run",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "llm_wrapper",
-        "output_processor": ["metamodel/maps/attributes/inference/haystack_entities.json"]
-    },
-    {
-        "package": "haystack.core.pipeline.pipeline",
-        "object": "Pipeline",
-        "method": "run",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "task_wrapper"
-    }
-] 
-}
