--- src/monocle_apptrace/metamodel/maps/llamaindex_methods.json
@@ -1,92 +1,0 @@
-{
-"wrapper_methods" : [
-    {
-        "package": "llama_index.core.indices.base_retriever",
-        "object": "BaseRetriever",
-        "method": "retrieve",
-        "span_name": "llamaindex.retrieve",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "task_wrapper",
-        "output_processor": ["metamodel/maps/attributes/retrieval/llamaindex_entities.json"]
-    },
-    {
-        "package": "llama_index.core.indices.base_retriever",
-        "object": "BaseRetriever",
-        "method": "aretrieve",
-        "span_name": "llamaindex.retrieve",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "atask_wrapper",
-        "output_processor": ["metamodel/maps/attributes/retrieval/llamaindex_entities.json"]
-    },
-    {
-        "package": "llama_index.core.base.base_query_engine",
-        "object": "BaseQueryEngine",
-        "method": "query",
-        "span_name": "llamaindex.query",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "task_wrapper"
-    },
-    {
-        "package": "llama_index.core.base.base_query_engine",
-        "object": "BaseQueryEngine",
-        "method": "aquery",
-        "span_name": "llamaindex.query",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "atask_wrapper"
-    },
-    {
-        "package": "llama_index.core.llms.custom",
-        "object": "CustomLLM",
-        "method": "chat",
-        "span_name": "llamaindex.llmchat",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "task_wrapper",
-        "output_processor": ["metamodel/maps/attributes/inference/llamaindex_entities.json"]
-    },
-    {
-        "package": "llama_index.core.llms.custom",
-        "object": "CustomLLM",
-        "method": "achat",
-        "span_name": "llamaindex.llmchat",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "atask_wrapper",
-        "output_processor": ["metamodel/maps/attributes/inference/llamaindex_entities.json"]
-    },
-    {
-        "package": "llama_index.llms.openai.base",
-        "object": "OpenAI",
-        "method": "chat",
-        "span_name": "llamaindex.openai",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "llm_wrapper",
-        "output_processor": ["metamodel/maps/attributes/inference/llamaindex_entities.json"]
-    },
-    {
-        "package": "llama_index.llms.openai.base",
-        "object": "OpenAI",
-        "method": "achat",
-        "span_name": "llamaindex.openai",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "allm_wrapper",
-        "output_processor": ["metamodel/maps/attributes/inference/llamaindex_entities.json"]
-    },
-    {
-        "package": "llama_index.llms.mistralai.base",
-        "object": "MistralAI",
-        "method": "chat",
-        "span_name": "llamaindex.mistralai",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "llm_wrapper",
-        "output_processor": ["metamodel/maps/attributes/inference/llamaindex_entities.json"]
-    },
-    {
-        "package": "llama_index.llms.mistralai.base",
-        "object": "MistralAI",
-        "method": "achat",
-        "span_name": "llamaindex.mistralai",
-        "wrapper_package": "wrap_common",
-        "wrapper_method": "allm_wrapper",
-        "output_processor": ["metamodel/maps/attributes/inference/llamaindex_entities.json"]
-    }
-]
-}
