--- src/monocle_apptrace/metamodel/spans/span_types.json
@@ -1,16 +1,0 @@
-{
-    "span_types" : [
-        {
-            "type": "inference",
-            "description": "Model inference span"
-        },
-        {
-            "type": "retrieval",
-            "description": "vector embedding retrieval"
-        },
-        {
-            "type": "workflow",
-            "description": "workflow orchetraction at top level"
-        }
-    ]
-}
